+++ src/main/java/com/yimei/finance/controllers/admin/restfulapi/company/CompanyController.java
@@ -24,7 +24,7 @@ public class CompanyController {
 
     @RequestMapping(value = "/{id}", method = RequestMethod.GET)
     @ApiOperation(value = "根据id获取公司对象", response = CompanyObject.class)
-    @ApiImplicitParam(name = "id", value = "公司 id", dataType = "Long", paramType = "path")
+    @ApiImplicitParam(name = "id", value = "公司 id", dataType = "int", paramType = "path")
     public Result findByIdMethod(@PathVariable("id")Long id) {
         return companyService.findByIdWithAuthority(id, adminSession.getUser().getId());
     }
@@ -60,7 +60,7 @@ public class CompanyController {
 
     @RequestMapping(value = "/{id}", method = RequestMethod.PUT)
     @ApiOperation(value = "修改公司", notes = "根据 company 对象修改用户", response = CompanyObject.class)
-    @ApiImplicitParam(name = "id", value = "公司 id", dataType = "Long", paramType = "path")
+    @ApiImplicitParam(name = "id", value = "公司 id", dataType = "int", paramType = "path")
     public Result editCompanyMethod(@PathVariable("id") Long id,
                                     @ApiParam(name = "company", value = "公司对象", required = true) @Validated(value = {EditCompany.class}) @RequestBody CompanyObject company) {
         return companyService.editCompany(id, company, adminSession.getUser().getId());
@@ -68,7 +68,7 @@ public class CompanyController {
 
     @RequestMapping(value = "/{id}", method = RequestMethod.DELETE)
     @ApiOperation(value = "删除公司", notes = "根据 id 删除公司", response = CompanyObject.class)
-    @ApiImplicitParam(name = "id", value = "公司 id", dataType = "Long", paramType = "path")
+    @ApiImplicitParam(name = "id", value = "公司 id", dataType = "int", paramType = "path")
     public Result deleteCompanyMethod(@PathVariable("id") Long id) {
         return companyService.deleteCompany(id, adminSession.getUser().getId());
     }
